refactor: button 구조 개선
