+++ libs/shared/ui-foundation/src/libs/BaseButton/index.ts
@@ -1,0 +1,2 @@
+export { BaseButton } from './BaseButton';
+export type { ButtonProps, BaseButtonKind } from './BaseButton';
