+++ libs/shared/ui-foundation/src/libs/ContainedButton/ContainedButton.tsx
@@ -1,36 +1,12 @@
-import { Button as BaseButton } from '@kuma-ui/core';
-import { ButtonHTMLAttributes, ReactNode } from 'react';
+import { BaseButton, BaseButtonKind, ButtonProps } from '../BaseButton';
 
-type ButtonProps = ButtonHTMLAttributes<HTMLButtonElement> & {
-  children: ReactNode;
-  kind: keyof typeof kindPreset;
+type ContainedButtonProps = ButtonProps & {
+  kind?: Exclude<BaseButtonKind, 'text'>;
 };
 
-const kindPreset = {
-  text: {
-    bgColor: 'transparent',
-    textColor: 'colors.dark.900',
-  },
-  dark: {
-    bgColor: 'colors.dark.900',
-    textColor: '#fff',
-  },
-};
-
-export const ContainedButton = ({ children, kind, ...props }: ButtonProps) => {
+export const ContainedButton = ({ children, kind = 'dark', ...props }: ContainedButtonProps) => {
   return (
-    <BaseButton
-      px={14}
-      py={8}
-      bg={kindPreset[kind].bgColor ?? '#ff0000'}
-      color={kindPreset[kind].textColor ?? '#fff'}
-      fontSize={14}
-      fontWeight={'fontWeights.medium'}
-      borderRadius={12}
-      border="1px solid"
-      borderColor={kindPreset[kind].bgColor ?? '#ff0000'}
-      {...props}
-    >
+    <BaseButton kind={kind} {...props}>
       {children}
     </BaseButton>
   );
